Adds 'Park Now!' button and updates database upon click
I added a button to the map, which reads 'Park Now!'. When clicked,
the event handler gets the current location of the 'current location'
marker on the screen, and sends a post request to create a new
parking_spot record in the database. Next step is get all parking_spots
of a user from the database, rendering the current parking spot
differently from past parking spots, and allowing the user to toggle
showing past parking spots or not.