Adds date/time, note, and remove button to button drawer
Now, instead of a modal opening up when a user clicks on
a past parking location, the bottom drawer opens up showing
the date and time when the user parked, a note if they
provided one, and a button they can click to remove
the parking record from the database. Next I need to clear out
the html from the drawer upon delete the record, and I need
to implement a similar functionality for the current parking location.