Moves newParkingSpot action into bottom drawer
Previously, when a user clicked on the user icon,
a modal popped up asking if they wanted to save a new
parking spot. I think that took away from the fluid user
experience, so now, when a user clicks the user icon, it
automatically saves that parking spot into the database,
and then displays info about that parking spot in the bottom
drawer.  I user can then add a note, or immediately delete that
parking record.